requirements.txt
pymongo
streamlit
python-dotenv
pandas
